fix: decouple api token table
