--- frontend/src/component/admin/api/index.tsx
@@ -1,11 +1,0 @@
-import { ApiTokenPage } from 'component/admin/apiToken/ApiTokenPage/ApiTokenPage';
-
-export const ApiPage = () => {
-    return (
-        <div>
-            <ApiTokenPage />
-        </div>
-    );
-};
-
-export default ApiPage;
