+++ frontend/src/component/common/ApiTokenTable/CreateApiTokenButton/CreateApiTokenButton.tsx
@@ -8,20 +8,23 @@ import { useNavigate } from 'react-router-dom';
 import { Add } from '@mui/icons-material';
 import { useOptionalPathParam } from 'hooks/useOptionalPathParam';
 
-export const CreateApiTokenButton = () => {
+interface ICreateApiTokenButton {
+    path: string;
+    permission: string;
+    project?: string;
+}
+
+export const CreateApiTokenButton = ({
+    path,
+    permission,
+    project,
+}: ICreateApiTokenButton) => {
     const navigate = useNavigate();
-
-    const project = useOptionalPathParam('projectId');
-
-    const to = Boolean(project) ? 'create' : '/admin/api/create-token';
-    const permission = Boolean(project)
-        ? CREATE_PROJECT_API_TOKEN
-        : CREATE_API_TOKEN;
 
     return (
         <ResponsiveButton
             Icon={Add}
-            onClick={() => navigate(to)}
+            onClick={() => navigate(path)}
             data-testid={CREATE_API_TOKEN_BUTTON}
             permission={permission}
             projectId={project}
